Add files via upload
This is an initial cut of the toy BiggestPrimeIn90s. The project intends to calculate a prime number that is as big as possible in 60 seconds. This initial cut contains an initial design framework and a couple implementation of test calculator and display classes. The current design aims at allowing different, possible concurrent calculators and different ways of displaying a timer and the calculated prime number.