Add files via upload
Check in a more serious attempt to get a prime number. It utilizes results related to the Mersenne conjecture. The algorithm is a simple parallel algorithm. Not tested as I was programming on an old installation of VS 2008. Time is running out though and I will make this my final check-in for interview purpose.